change: move password validation to /src/utils
